requirements.txt
requests==2.32.5
python-telegram-bot==21.6
python-dotenv==1.0.1
customtkinter==5.2.2